Login n Reg

+++ php/login.php
@@ -12,13 +12,16 @@ if (isset($_REQUEST['signup_email'])) $email = sanitizeString($_REQUEST['signup_
 if ($email === "") die('1');
 if (isset($_REQUEST['signup_password'])) $pass = sanitizeString($_REQUEST['signup_password']);
 if ($pass === "") die('2');
-$sql="SELECT `email` FROM `users` WHERE `email`='$email' AND `pass`='$pass'";
+$sql="SELECT * FROM `users` WHERE `email`='$email' AND `pass`='$pass'";
 $result=sql($sql);
 if($result->num_rows>0){
-    $token=randomString(64);
-    sql("DELETE FROM `cookie` WHERE 'email'='$email'");
-    $result=sql("INSERT INTO `cookie` (`mail`,`token`) VALUES ('$email','$token')");
-    $_COOKIE['convo_mail']=$email;
-    $_COOKIE['convo_token']=$token;
-    $_SESSION['on']='1';
+    $row=$result->fetch_assoc();
+    if($row['confirmation']!='0') {
+        $token = randomString(64);
+        sql("DELETE FROM `cookie` WHERE 'email'='$email'");
+        $result = sql("INSERT INTO `cookie` (`mail`,`token`) VALUES ('$email','$token')");
+        $_COOKIE['convo_mail'] = $email;
+        $_COOKIE['convo_token'] = $token;
+        $_SESSION['on'] = '1';
+    }else die('-1');
 }else die('0');
